front: split catalog import from the old itinerary modal
