+++ front/src/applications/operationalStudies/views/Scenario/components/Timetable/TimetableToolbar.tsx
@@ -17,17 +17,23 @@ import cx from 'classnames';
 import { useTranslation } from 'react-i18next';
 
 import { useScenarioContext } from 'applications/operationalStudies/hooks/useScenarioContext';
+import useScenarioTrainScheduleSet, {
+  type ImportTrainScheduleSetsPayload,
+} from 'applications/operationalStudies/hooks/useScenarioTrainScheduleSet';
 import { useTimetableContext } from 'applications/operationalStudies/hooks/useTimetableContext';
 import { osrdEditoastApi } from 'common/api/osrdEditoastApi';
 import { ModalContext } from 'common/BootstrapSNCF/ModalSNCF/ModalProvider';
 import MenuTriggerButton from 'common/MenuTriggerButton';
 import UploadFileModal from 'common/uploadFileModal';
 import type { TrainScheduleWithDetails } from 'modules/trainSchedule/types';
+import { setFailure } from 'reducers/main';
 import { resetItineraryForm } from 'reducers/osrdconf/operationalStudiesConf';
 import { useAppDispatch } from 'store';
+import { castErrorToFailure } from 'utils/error';
 
 import { MANAGE_TRAIN_SCHEDULE_TYPES } from '../../consts';
 import useImportTrainSchedules from '../ImportTrainSchedule';
+import { TrainScheduleSetCatalogDialog } from '../ImportTrainScheduleSets';
 import RoundTripsModal from '../RoundTrips/RoundTripsModal';
 import FilterPanel from './FilterPanel';
 import SelectionToolBar from './TimetableSelectionToolbar';
@@ -36,6 +42,7 @@ import { exportTrainSchedules, timetableHasInvalidTrainSchedule } from './utils'
 
 type TimetableToolbarProps = {
   timetableFilters: TimetableFilters;
+  trainSchedulesWithDetails: TrainScheduleWithDetails[];
   filteredTrainSchedules: TrainScheduleWithDetails[];
   selectedTrainScheduleIds: number[];
   showTrainDetails: boolean;
@@ -53,6 +60,7 @@ type TimetableToolbarProps = {
 
 const TimetableToolbar = ({
   timetableFilters,
+  trainSchedulesWithDetails,
   filteredTrainSchedules,
   selectedTrainScheduleIds,
   showTrainDetails,
@@ -128,6 +136,19 @@ const TimetableToolbar = ({
       />
     );
 
+  const { importTrainScheduleSets } = useScenarioTrainScheduleSet(trainSchedulesWithDetails);
+  const [trainScheduleSetsCatalogModalIsOpen, setTrainScheduleSetsCatalogModalIsOpen] =
+    useState<boolean>(false);
+
+  const handleImportTrainScheduleSets = async (data: ImportTrainScheduleSetsPayload) => {
+    try {
+      await importTrainScheduleSets(data);
+      setTrainScheduleSetsCatalogModalIsOpen(false);
+    } catch (error) {
+      dispatch(setFailure(castErrorToFailure(error)));
+    }
+  };
+
   return (
     <>
       {areInvalidTrainSchedules && (
@@ -197,8 +218,7 @@ const TimetableToolbar = ({
                   icon: <Book />,
                   title: t('timetable.importTimetableNetworkGraphFromCatalog'),
                   dataTestID: 'scenarios-import-timetable-by-catalog',
-                  onClick: () =>
-                    setDisplayTrainScheduleManagement(MANAGE_TRAIN_SCHEDULE_TYPES.catalog),
+                  onClick: () => setTrainScheduleSetsCatalogModalIsOpen(true),
                 },
               ],
             }}
@@ -285,6 +305,12 @@ const TimetableToolbar = ({
           timetableId={timetableId}
           trainSchedules={trainSchedules}
           refreshNge={refreshNge}
+        />
+      )}
+      {trainScheduleSetsCatalogModalIsOpen && (
+        <TrainScheduleSetCatalogDialog
+          onSubmit={handleImportTrainScheduleSets}
+          onCancel={() => setTrainScheduleSetsCatalogModalIsOpen(false)}
         />
       )}
     </>
