+++ front/src/applications/operationalStudies/views/Scenario/consts.ts
@@ -2,6 +2,5 @@ export const MANAGE_TRAIN_SCHEDULE_TYPES = Object.freeze({
   none: 'NONE',
   add: 'ADD',
   edit: 'EDIT',
-  catalog: 'CATALOG',
   itinerary: 'ITINERARY',
 });
